Organize docs under docs/ and unify server analysis folder
- Move authoritative Markdown into docs/ (architecture, deployment, web, AGENTS)
- Add stubs at repo root and web/ pointing to canonical paths
- Merge server/engine and server/scan into server/analysis; update imports and docs page
