+++ docs/README.md
@@ -1,0 +1,12 @@
+# Monix documentation
+
+Authoritative Markdown for the repository lives here.
+
+| Document | Description |
+| --- | --- |
+| [ARCHITECTURE.md](./ARCHITECTURE.md) | System design, repository layout, module boundaries |
+| [DEPLOYMENT.md](./DEPLOYMENT.md) | Production deployment and operations checklist |
+| [web.md](./web.md) | Next.js app (`web/`): setup, env vars, testing |
+| [AGENTS.md](./AGENTS.md) | Next.js Cursor agents index (generated metadata for AI-assisted work) |
+
+The repository root keeps short [`README.md`](../README.md) stubs that point here where files were consolidated.
