+++ web/src/app/api/dashboard/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { dashboardPayload } from "@/server/engine/dashboard-payload";
+import { dashboardPayload } from "@/server/analysis/dashboard-payload";
 
 export async function GET() {
   return NextResponse.json(dashboardPayload());
