+++ web/src/app/api/system-stats/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { getSystemStatsPayload } from "@/server/engine/system-stats";
+import { getSystemStatsPayload } from "@/server/analysis/system-stats";
 
 export async function GET() {
   return NextResponse.json({ status: "success", ...getSystemStatsPayload() });
