+++ web/src/app/api/threat-info/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { threatInfoPayload } from "@/server/engine/threat-catalog";
+import { threatInfoPayload } from "@/server/analysis/threat-catalog";
 
 export async function GET() {
   return NextResponse.json(threatInfoPayload());
